[AM][OPS-9321] move income form to AffordabilityForm
